Clean up and added WAcsXmlAccessor
Allows us to get data directly from WACS
