+++ package.json
@@ -10,6 +10,7 @@
     "@mui/material": "^5.15.15",
     "@vitejs/plugin-react": "^4.2.1",
     "axios": "^1.6.8",
+    "fast-xml-parser": "^4.3.6",
     "install": "^0.13.0",
     "proskomma": "^0.9.3",
     "react-dom": "^18.3.1",
