+++ src/api/SourceTextAccessor/Door43UsfmAccessor.ts
@@ -1,0 +1,157 @@
+import { books } from "src/applicationLogic/data/newTestamentMetadata";
+import { AlignedText, AlignedVerse, GreekAlignmentData } from "src/types";
+import { SourceTextAccessor, SourceTextFetcher, SourceTextResource } from ".";
+import { Proskomma } from "proskomma";
+
+
+export class Door43UsfmAccessor implements SourceTextAccessor, SourceTextFetcher {
+    sourceTextResource: SourceTextResource;
+    pk = new Proskomma();
+
+    constructor(resourceType: string, resourceLanguage: string) {    
+        this.sourceTextResource = {
+            resourceType: resourceType.toLowerCase(),
+            resourceLanguage: resourceLanguage.toLowerCase(),
+            resourceBaseURL: 
+                `https://git.door43.org/Door43-Catalog/${resourceLanguage}_${resourceType}/raw/branch/master/`,
+        };
+    }
+
+
+    setSourceTextResource(resourceType: string, resourceLanguage: string): void {
+        this.sourceTextResource = {
+            resourceLanguage: resourceType,
+            resourceType: resourceLanguage,
+            resourceBaseURL: 
+                `https://git.door43.org/Door43-Catalog/${resourceLanguage}_${resourceType}/raw/branch/master/`,
+        }
+    }
+
+
+    async getSourceText(bookName: string, chapter: number | string): Promise<AlignedVerse[]> {
+
+        let sourceText = await this.fetchSourceText(bookName)
+        let chapterNum = typeof(chapter) === "number" ? chapter : parseInt(chapter);
+
+        let alignedVerses : AlignedVerse[]= []
+        if(sourceText) {
+            this.setUpProskomma(sourceText);
+            let dataQuery = this.getDataQuery(chapter);
+            let usfmData = JSON.stringify(await this.getUSFMData(dataQuery));
+            alignedVerses = this.parseSourceText(usfmData, chapterNum);
+        }
+
+        return alignedVerses;
+    }
+
+
+    async fetchSourceText(bookName: string) : Promise<string | undefined> {
+        try {
+            let book;
+            book = await fetch(`${this.sourceTextResource.resourceBaseURL}${books[bookName].abbreviatedBook}.usfm`);
+            return book.text()
+
+        } catch (error) {
+            return undefined;
+        }
+    }
+
+
+    private async setUpProskomma(bookContent: string) {
+        const mutation = `mutation { addDocument(` +
+        `selectors: [{key: "lang", value: "eng"}, {key: "abbr", value: "${this.sourceTextResource.resourceType}"}], ` +
+        `contentType: "usfm", ` +
+        `content: """${bookContent}""") }`;
+    
+        await this.pk.gqlQuery(mutation);
+    }
+
+
+    private getDataQuery(chapter: string | number) {
+        return `{
+            documents {
+                id
+                cv(chapter: "${chapter}") {
+                    items {
+                        subType
+                        payload
+                    }
+                }
+            }
+        }`
+    }
+
+
+    private async getUSFMData(dataQuery: string) {
+        const result = await this.pk.gqlQuery(dataQuery);
+        let cvData = result?.data?.documents[0]?.cv[0]?.items.filter((item: any) => 
+            item.payload === "milestone/zaln" || item.subType === "wordLike" 
+            || (item.payload.includes("x-strong") && item.subType === "start")
+            || item.payload.includes("verse/")
+        )
+        return cvData; 
+    }
+
+
+    parseSourceText(sourceText: string, chapter: number): AlignedVerse[] {
+        let alignedVerses : AlignedVerse[] = [];
+        let alignedText : AlignedText[] = [];
+        let alignStartCount = 0; 
+        let text = "";
+        let greekAlignmentData : GreekAlignmentData[] = []
+        let verseNum = 1;
+
+        let cvData = JSON.parse(sourceText);
+    
+        for(let i = 0; i < cvData.length; i++) {
+            let attribute = cvData[i];
+    
+            if(attribute.subType === "start" && attribute.payload === "milestone/zaln") {
+                alignStartCount++; 
+            } else if(attribute.subType === "end" && attribute.payload === "milestone/zaln") {
+                alignStartCount--; 
+            } else if(attribute.subType === "wordLike") {
+                let startingChar = text ===   "" ? "" : " "
+                text += startingChar + attribute.payload;
+            } else if(attribute.subType === "start" && attribute.payload.includes("x-strong")) {
+                let strong = this.getStrongs(attribute.payload);
+                if(strong) {
+                    greekAlignmentData.push({strong: strong});
+                }
+            } else if(attribute.subType === "end" && attribute.payload.includes("verse/")) {
+                alignedVerses.push({verseNum: verseNum, alignedVerseText: alignedText})
+                verseNum++;
+                alignedText = []
+            }
+    
+            if(alignStartCount === 0 && text !== "") {
+                // Push alignment text
+                if(greekAlignmentData.length > 0) {
+                    alignedText.push({text: text, greekAlignmentData: greekAlignmentData});
+                } else {
+                    alignedText.push({text: text});
+                }
+    
+                // reset buffers
+                text = "";
+                greekAlignmentData = [];
+            }
+        }
+        return alignedVerses;
+    }
+    
+
+    private getStrongs(str: string) {
+        const regex = /\b[Gg]\d+\b/;
+        const match = str.match(regex);
+        
+        if (match) {
+          const extracted = match[0];
+          if(extracted.length > 4 && extracted.charAt(extracted.length - 1) === "0") {
+            return extracted.substring(0, extracted.length - 1)
+          } else {
+            return undefined;
+          }
+        }
+    }
+}
